Refactor Form to accept external schema config
