+++ web/src/components/viavai/Form.tsx
@@ -9,41 +9,6 @@ import {
     type FieldDefinition,
 } from '@/types/viavai/form.types';
 
-/* ============================================================
-   JSON Form Configuration
-============================================================ */
-
-const sample: Component[] = [
-    {
-        type: 'text',
-        name: 'title',
-        label: 'Bug Title',
-        description: 'Short summary of the issue.',
-        placeholder: 'Login button not working',
-        required: true,
-        default: '',
-        validate: {
-            minLength: 5,
-            maxLength: 32,
-        },
-        error: 'Title must be between 5 and 32 characters.',
-    },
-    {
-        type: 'textarea',
-        name: 'description',
-        label: 'Description',
-        description: 'Steps to reproduce and expected result.',
-        placeholder: 'Explain what happened...',
-        required: true,
-        default: '',
-        validate: {
-            minLength: 20,
-            maxLength: 100,
-        },
-        error: 'Description must be between 20 and 100 characters.',
-    },
-];
-
 // Registry
 import { textField } from '@/components/fields/Text';
 import { emailField } from '@/components/fields/Email';
@@ -70,13 +35,17 @@ function buildSchema(components: Component[]) {
 }
 
 // Component
-export function Form() {
-    const schema = buildSchema(sample);
+type FormProps = {
+    schema: Component[];
+};
+
+export function Form({ schema: components }: FormProps) {
+    const schema = buildSchema(components);
 
     const form = useForm({
         resolver: zodResolver(schema),
         defaultValues: Object.fromEntries(
-            sample.map((f) => [f.name, f.default ?? ''])
+            components.map((f) => [f.name, f.default ?? ''])
         ),
     });
 
@@ -96,7 +65,7 @@ export function Form() {
             onSubmit={form.handleSubmit(onSubmit)}
         >
             <FieldGroup>
-                {sample.map((config) => {
+                {components.map((config) => {
                     const definition = fieldRegistry[config.type];
                     if (!definition) return null;
